Blog debugging session complete
This is a stable push for the blog debugging session. I need to correct some issues with the layout. See the original template found within the BlogPost folder within the Views directory for the notes regarding what went wrong with the layout.

Make changes to the Blog Upload Image functionality. Follow the intial documentation and confirm the process was followed correctly. Currently there is an issue that needs to be resolved regarding the functionality when a null parameter is passed through the image uploader.